Add ability to specify a Stylesheet when rendering to reStructuredText
Rendering a note written in reStructuredText is a nice feature but
the default stylesheet, the one included in docutils, is not likely to
suit everyone's needs.  This commit enables the user to specify a
stylesheet to use when rendering the reST formatted note.  It adds one
new configuration parameter, rest_css_path, the full path to the
user's stylesheet.  The stylesheet is embedded into the reST-rendered
page.  Customizing the stylesheet allows users to do customizations
that enhance the notes.  My motivation for this change was to enable
the use of 'strikethrough' text on list items that have been
completed.