UI makes clear when no not is currently selected
When no note is selected, either because the search string is too
strict, or because the user has clicked past the end of the notes list,
the user interface now makes it clear that no note has been selected,
and suggests some alternative actions to the user.

Thank you to Christian Tietze for reporting the no-note-selected issues.