Correcting outdated install instructions
Updating documentation to clarify current requirements as per https://github.com/cpbotha/nvpy/issues/190 . Also, on Debian and other distros which still maintain a 2.x as the default `python`, `pip install` will result in a broken installation. I believe `pip3` is generally aliased  on any system with a 3.x release so I think this is a more reasonable default instruction?